+++ README.md
@@ -29,12 +29,14 @@
 - [x] 게임 상의 오류 메시지를 웹으로 전송해 출력한다.
 
 #### 5단계 - DB 적용
-- [ ] 체스 진행 정보를 데이터베이스에 저장하고 불러 올 수 있다. 
+- [x] 체스 진행 정보를 데이터베이스에 저장하고 불러 올 수 있다. 
   - [x] 데이터베이스(MySQL)에 연결한다.
   - [x] 게임 정보를 저장/업데이트/불러오기 할 수 있다. 
   - [x] 체스판 정보를 저장/업데이트/불러오기 할 수 있다.
-  - [x] Move 성공 시 게임, 체스판 정보를 저장/업데이트한다. 
-- [ ] 서버를 재시작하더라도 이전에 하던 체스 게임을 이어서 진행할 수 있다. 
+  - [x] Move 성공 시 게임, 체스판 정보를 저장/업데이트한다.
+  - [x] End 성공 시 체스판 정보를 업데이트한다.
+  - [x] Start 성공 시 게임, 체스판 정보를 새롭게 저장한다.
+- [x] 서버를 재시작하더라도 이전에 하던 체스 게임을 이어서 진행할 수 있다. 
   - [x] 이전에 하던 체스 게임을 불러와 웹 UI에 출력한다.
 
 #### 더 해보기
