+++ src/main/java/chess/service/ChessService.java
@@ -8,11 +8,9 @@ import chess.domain.game.state.Ended;
 import chess.domain.game.state.Ready;
 import chess.domain.game.state.Started;
 import chess.domain.game.state.State;
-import chess.domain.piece.BishopPiece;
 import chess.domain.piece.Color;
 import chess.domain.position.Position;
 import chess.dto.BoardDto;
-import chess.dto.ChessResponseDto;
 import chess.dto.GameDto;
 import chess.dto.MoveRequestDto;
 import chess.dto.PieceDto;
@@ -48,28 +46,37 @@ public class ChessService {
         return new Ended(board);
     }
 
-    public void move(final ChessGame chessGame, final MoveRequestDto moveDto) {
-        try {
-            final Position from = Position.create(moveDto.getSource());
-            final Position to = Position.create(moveDto.getTarget());
-            chessGame.move(from, to);
+    public void start(final ChessGame chessGame) {
+        initializeChessGame(chessGame);
+        chessGame.start();
 
-            final Integer id = gameDao.findMaxId();
-            boardDao.updateOnePosition(id, from.getName(), new PieceDto(chessGame.getBoard().get(from)));
-            boardDao.updateOnePosition(id, to.getName(), new PieceDto(chessGame.getBoard().get(to)));
+        gameDao.save(GameDto.of(chessGame));
+        final Integer id = gameDao.findMaxId();
+        boardDao.save(BoardDto.of(id, chessGame.getBoard()));
+    }
 
-            gameDao.updateById(GameDto.of(id, chessGame.getState(), chessGame.getState().getTurn()));
-        } catch (final Exception e) {
-            throw new IllegalArgumentException(e.getMessage());
+    private void initializeChessGame(final ChessGame chessGame) {
+        if (!chessGame.isNotEnded()) {
+            chessGame.initialize();
         }
     }
 
-    private ChessResponseDto move1(final ChessGame chessGame, final MoveRequestDto moveDto) {
-        try {
-            chessGame.move(Position.create(moveDto.getSource()), Position.create(moveDto.getTarget()));
-            return new ChessResponseDto(chessGame);
-        } catch (final Exception e) {
-            return new ChessResponseDto("error", e.getMessage(), chessGame);
-        }
+    public void move(final ChessGame chessGame, final MoveRequestDto moveDto) {
+        final Position from = Position.create(moveDto.getSource());
+        final Position to = Position.create(moveDto.getTarget());
+        chessGame.move(from, to);
+
+        final Integer id = gameDao.findMaxId();
+        boardDao.updateOnePosition(id, from.getName(), new PieceDto(chessGame.getBoard().get(from)));
+        boardDao.updateOnePosition(id, to.getName(), new PieceDto(chessGame.getBoard().get(to)));
+
+        gameDao.updateById(GameDto.of(id, chessGame.getState(), chessGame.getState().getTurn()));
+    }
+
+    public void end(final ChessGame chessGame) {
+        chessGame.end();
+
+        final Integer id = gameDao.findMaxId();
+        gameDao.updateById(GameDto.of(id, chessGame.getState(), chessGame.getState().getTurn()));
     }
 }
